refactor(package): divide business logic into internal packages
According to [golang-standards/project-layout](https://github.com/golang-standards/project-layout/tree/master/cmd),
it's best to have a small main function that imports methods from /pkg (public library code) or /internal (private) packages.
Let's move all of kleat's business logic into internal packages for now, and we can decide what precisely should be
exportable (for instance, for the operator) as needed.
