+++ internal/validate/validate_hal.go
@@ -1,0 +1,70 @@
+/*
+ * Copyright The Spinnaker Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package validate
+
+import (
+	"errors"
+	"github.com/spinnaker/kleat/api/client"
+	"strings"
+)
+
+type ValidationFailure struct {
+	msg string
+}
+
+type HalConfigValidator func(*client.HalConfig) []ValidationFailure
+
+func getValidators() []HalConfigValidator {
+	return []HalConfigValidator{
+		validateKindsAndOmitKinds,
+	}
+}
+
+func ValidateHalConfig(h *client.HalConfig) error {
+	messages := getValidationMessages(h, getValidators())
+	if len(messages) > 0 {
+		msg := strings.Join(messages, "\n")
+		return errors.New(msg)
+	}
+	return nil
+}
+
+func getValidationMessages(h *client.HalConfig, fa []HalConfigValidator) []string {
+	var messages []string
+	for _, f := range fa {
+		rs := f(h)
+		for _, r := range rs {
+			messages = append(messages, r.msg)
+		}
+	}
+	return messages
+}
+
+func validateKindsAndOmitKinds(h *client.HalConfig) []ValidationFailure {
+	var messages []ValidationFailure
+	for _, a := range h.Providers.Kubernetes.Accounts {
+		if !(len(a.Kinds) == 0) && !(len(a.OmitKinds) == 0) {
+			messages = append(messages, fatalResult("Cannot specify both kinds and omitKinds."))
+		}
+	}
+	return messages
+}
+
+func fatalResult(msg string) ValidationFailure {
+	return ValidationFailure{
+		msg: msg,
+	}
+}
